Update README with typo fixes
As part of the initiative to promote open source culture at Intuit, we are reviewing the text of projects featured on OpenSource.Intuit.com for spelling and grammar. Thanks!As part of the initiative to promote open source culture at Intuit, we are reviewing the text of projects featured on OpenSource.Intuit.com for spelling and grammar. Thanks!